rename: 컴포넌트 폴더 이름 및 구조 컨벤션으로 수정
