+++ src/app/components/Tab/panel/TabPanel.module.scss
